% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{bivariate}
\alias{bivariate}
\title{Make the bivariate report.}
\usage{
bivariate(x, target, targetClass = c("binary", "numeric"))
}
\arguments{
\item{x}{Vector of explanatory variable.}

\item{target}{Vector of dependent variable.}

\item{targetClass}{Tells whether the dependent variable is numeric or binary.}
}
\value{
Returns a tibble.
}
\description{
Receives x (explanatory variable) and y (dependent variable) and
 calculates bivariate metrics according to the types of the input variables.
}
\seealso{
descriptive_numeric, descriptive_categoric
}
