% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_all.R
\name{bivariate_all}
\alias{bivariate_all}
\title{Make the bivariate report for all dataset.}
\usage{
bivariate_all(
  data,
  target,
  targetClass = c("binary", "numeric"),
  cutNum = TRUE,
  cutNumMethod = c("median", "quintile", "decile"),
  labelNA = "Missing",
  export = FALSE,
  NameExport = "BivariateReport"
)
}
\arguments{
\item{data}{data.frame or tibble.}

\item{target}{Vector of dependent variable.}

\item{targetClass}{Tells whether the dependent variable is numeric or binary.}

\item{cutNum}{Discretize the explanotory numeric variable? TRUE by default.}

\item{cutNumMethod}{Method for discretization if cutNum = TRUE. Disponible "median", "quintile" or "decile".}

\item{labelNA}{Label of missing values assume. "Missing" by default.}

\item{export}{TRUE of FALSE for export results in excel arquive .xlsx. FALSE by default.}

\item{NameExport}{Character with name of the excel arquive .xlsx file to be exported. By default the is BivariateReport.}
}
\value{
Returns a tibble.
}
\description{
Receives a data.frame(or tibble) and y (dependent variable) and
 calculates bivariate metrics according to the types of the input variables ind dataset.
}
\seealso{
descriptive_numeric, descriptive_categoric, bivariate
}
