% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_categoric.R
\name{descriptive_categoric}
\alias{descriptive_categoric}
\title{Make the descriptive statistical analysis of categoric variables}
\usage{
descriptive_categoric(
  data,
  export = FALSE,
  NameExport = "DescriptiveCategoric",
  labelNA = "Missing"
)
}
\arguments{
\item{data}{data.frame or tibble.}

\item{export}{TRUE of FALSE for export results in excel arquive .xlsx. FALSE by default.}

\item{NameExport}{Character with name of the excel arquive .xlsx file to be exported. By default the is DescritptiveCategoric.}
}
\value{
Returns a tibble.
}
\description{
Receives a data.frame or tibble and returns some descriptive measures for the categoric variables in the dataset.
}
\seealso{
descriptive_numeric
}
