% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_numeric.R
\name{descriptive_numeric}
\alias{descriptive_numeric}
\title{Make the descriptive statistical analysis of numeric variables}
\usage{
descriptive_numeric(data, export = FALSE, NameExport = "DescritptiveNumeric")
}
\arguments{
\item{data}{data.frame or tibble.}

\item{export}{TRUE of FALSE for export results in excel arquive .xlsx. FALSE by default.}

\item{NameExport}{Character with name of the excel arquive .xlsx file to be exported. By default the is DescritptiveNumeric.}
}
\value{
Returns a tibble.
}
\description{
Receives a data.frame or tibble and returns some descriptive measures for the numeric variables in the dataset.
}
\seealso{
descriptive_categoric
}
