% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformatioGain.R
\name{InformationGain}
\alias{InformationGain}
\title{Internal function for usage in bivariate function.}
\usage{
InformationGain(tble)
}
\description{
Internal function for usage in bivariate function.
}
\seealso{
bivariate
}
