% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{Internal function for usage in descriptive_categoric function.}
\usage{
univariate(x)
}
\description{
Internal function for usage in descriptive_categoric function.
}
\seealso{
descriptive_categoric
}
