% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{bivariate}
\alias{bivariate}
\title{Make the bivariate report.}
\usage{
bivariate(
  x,
  target,
  targetClass = c("binary", "numeric"),
  cutNum = TRUE,
  cutNumMethod = c("median", "quintile", "decile"),
  labelNA = "Missing"
)
}
\arguments{
\item{x}{Vector of explanatory variable.}

\item{target}{Vector of dependent variable.}

\item{targetClass}{Tells whether the dependent variable is numeric or binary.}

\item{cutNum}{Discretize the explanotory numeric variable? TRUE by default.}

\item{cutNumMethod}{Method for discretization if cutNum = TRUE. Disponible "median", "quintile" or "decile".}

\item{labelNA}{Label of missing values assume. "Missing" by default.}
}
\value{
Returns a tibble.
}
\description{
Receives x (explanatory variable) and y (dependent variable) and
 calculates bivariate metrics according to the types of the input variables.
}
\seealso{
descriptive_numeric, descriptive_categoric
}
